🔀 [BE] merge: from be/payment-charge/S12P21C209-153 to dev

See merge request s12-fintech-finance-sub1/S12P21C209!66

+++ back/payment/src/main/java/com/awoo/payment/application/QueryPaymentService.java
@@ -2,9 +2,7 @@ package com.awoo.payment.application;
 
 import com.awoo.payment.application.query.FetchBalanceQuery;
 
-import java.math.BigDecimal;
-
 public interface QueryPaymentService {
 
-    BigDecimal fetchBalance(FetchBalanceQuery query);
+    int fetchBalance(FetchBalanceQuery query);
 }
